Let TimeConstants functions accept doubles

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomousevent/TimeConstants.java
@@ -3,11 +3,11 @@ package org.firstinspires.ftc.teamcode.autonomousevent;
 public class TimeConstants {
     public static final int DEGREE_90 = 660;
 
-    public static int degreeTime(int degree){
+    public static int degreeTime(double degree){
         return 0;
     }
 
-    public static int distanceTime(int distance){
+    public static int distanceTime(double distance){
         return 0;
     }
